fix/ 비행선 공격력/체력 더블로 변경, 데미지인포도 더블로 변경

+++ Assets/Scripts/Enemy/TestScript/TestAirshipHealth.cs
@@ -25,7 +25,7 @@ public class TestAirshipHealth : MonoBehaviour, IDamageable
         DamageInfo damageInfo)
     {
         float damageAmount =
-            damageInfo.Damage;
+            (float)damageInfo.Damage;
 
         if (damageAmount <= 0f ||
             currentHealth <= 0)
